Add a devcontainer setup and initial build system
This commit configures the project to use the [DocOps
devcontainer][devcontainer], which provides a standard containerized
environment with pre-installed software for working on the book.

Currently, the DocOps devcontainer does not come with the `mdbook`
program pre-installed. I am happy to add `mdbook` to the devcontainer if
you think this is a valuable addition to the project. In the meantime, I
have added a `bootstrap.sh` script, which performs the necessary setup
to get `mdbook` working.

In addition, I added a Makefile to the project. At the moment, the only
thing the Makefile does is wrap the `mdbook` command. The Makefile fixes
the invocation of `mdbook serve` to integrate well with VS Code. The
Makefile also makes a minor enhancement to the `mdbook clean` command.

However, the primary function of the Makefile is to serve as a platform
for adding lint checks to the project. Follow-up commits will add these
lint checks one by one along with the necessary fixes for any errors
they identify.

I plan to lint the book's text for spelling, grammar, style, etc. The
DocOps devcontainer provides the software necessary to accomplish this.

[devcontainer]: https://github.com/doitintl/docops/#development-containers